Project ELEANOR ana kod dizisi/requirements.txt

RPi.GPIO
smbus2
INA219
requests
flask
